pyproject.toml
[project]
name = "mnist-pilot"
version = "0.1.0"
description = ""
authors = [
    {name = "hajime7227",email = "ser.hjm95272@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (>=2.0.0,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "streamlit (>=1.42.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
